% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_to_na.R
\name{complexn_to_na}
\alias{complexn_to_na}
\title{Replace \code{NaN} and/or \code{Inf} values}
\usage{
complexn_to_na(x, convert_inf = TRUE, convert_nan = TRUE, replace = NA)
}
\arguments{
\item{x}{a vector containing numeric values and real and imaginary parts of complex values}

\item{convert_inf}{logical indicating whether Inf values should be replaced - default to TRUE}

\item{convert_nan}{logical indicating whether NaN values should be replaced - default to TRUE}

\item{replace}{replacement value - defaults to NA}
}
\value{
vector with replaced \code{NA} values.
}
\description{
Replace \code{NaN} and/or \code{Inf} values
}
\examples{
x <- c(NaN, 1, 2, 3, Inf)
complexn_to_na(x)

}
