% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not in operator - courtesy of helfRlein}
\usage{
vec_match \%nin\% vec_compare
}
\arguments{
\item{vec_match}{vector of values to be matched}

\item{vec_compare}{vector of values to compare against}
}
\value{
boolean vector
}
\description{
Not in operator - courtesy of helfRlein
}
\examples{
c(1,2,3,4) \%nin\% c(1,2,5)

}
