% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_bar.R
\name{progress_bar}
\alias{progress_bar}
\title{Print a Progress Bar for For Loops - courtesy of helfRlein}
\usage{
progress_bar(run, max_run, width = 20L, info = run)
}
\arguments{
\item{run}{the iterator of the \code{for} loop or an integer with the current loop number.}

\item{max_run}{either an integer with the maximum number of loops if \code{run} is also a number,
or a vector with all possible iterations in the correct order.}

\item{width}{width an integer that indicates how wide the progress bar is printed.}

\item{info}{a string with additional information to be printed at the end of the line. The default is \code{run}.}
}
\description{
Print a Progress Bar for For Loops - courtesy of helfRlein
}
\examples{
for (i in 1:20) {
  Sys.sleep(0.1)
  progress_bar(run = i, max_run = 200, width = 60L)
}

}
