% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_list_na.R
\name{rm_na_list}
\alias{rm_na_list}
\title{Remove NAs from lists - courtesy of helfRlein}
\usage{
rm_na_list(x, recursive = FALSE)
}
\arguments{
\item{x}{object of type \code{list}}

\item{recursive}{logical indicating whether NAs contained within list elements should be removed too}
}
\value{
object of type \code{list} with NAs removed
}
\description{
Remove NAs from lists - courtesy of helfRlein
}
\examples{
x <- list(c(1:3), letters[1:4], NA, c(1, NA), list(c(5:6, NA), NA, "A"))
rm_na_list(x, recursive = TRUE)

}
